requirements.txt
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
colorzero==2.0
Flask==3.1.3
future==1.0.0
gpiozero==2.0.1
idna==3.11
iso8601==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
paho-mqtt==2.1.0
pyserial==3.5
PyYAML==6.0.3
requests==2.32.5
RPi.GPIO==0.7.1
setuptools==82.0.0
urllib3==2.6.3
Werkzeug==3.1.6
